requirements.txt
django >=4.1.1, <4.2.0
djangorestframework >=3.13.1, <3.14.0
pydantic >=1.10.2, <1.11.0